+++ src/features/routes/reflection-list/sidebar/Sidebar.tsx
@@ -6,8 +6,8 @@ import type { ReflectionTagCountList } from "@/src/api/reflection-api";
 import type { TagType } from "@/src/hooks/reflection-tag/useExtractTrueTags";
 import { NewFolderField } from "../side-folder-bar/NewFolderField";
 import { FolderItem, TagItem } from "./item";
+import { useFolder } from "@/src/hooks/folder/useFolder";
 import { tagMap } from "@/src/hooks/reflection-tag/useExtractTrueTags";
-import { useFolder } from "@/src/hooks/sidebar/folder/useFolder";
 import { theme } from "@/src/utils/theme";
 
 type SidebarProps = {
